+++ yash-env/CHANGELOG.md
@@ -37,6 +37,9 @@ A _private dependency_ is used internally and not visible to downstream users.
 - The `system::Read::read` method is now async. The `VirtualSystem`
   implementation of this method will block until data can be read from the file
   descriptor.
+- The `system::Write::write` method is now async. The `VirtualSystem`
+  implementation of this method will block until data can be written to the file
+  descriptor.
 - The `Env::get_tty` method is now async.
 - Private dependency versions:
     - derive_more 2.0.1 → 2.1.0